fix: revert "Changed size variable from INTEGER to BIGINT"
This reverts commit 646b19af731b8a25678d1b4bc0f378c3810783f4.

+++ src/syscheckd/src/db/src/fimDB.hpp
@@ -46,7 +46,7 @@ constexpr auto CREATE_FILE_DB_STATEMENT
     checksum TEXT NOT NULL,
     dev INTEGER,
     inode INTEGER,
-    size BIGINT,
+    size INTEGER,
     perm TEXT,
     attributes TEXT,
     uid TEXT,
